+++ MinecraftClient/MinecraftClient.csproj
@@ -75,6 +75,7 @@
     <Compile Include="AutoTimeout.cs" />
     <Compile Include="ChatBots\AutoDrop.cs" />
     <Compile Include="ChatBots\Mailer.cs" />
+    <Compile Include="Mapping\CubeFromWorld.cs" />
     <Compile Include="Mapping\Material2Tool.cs" />
     <Compile Include="ChatBots\MineCube.cs" />
     <Compile Include="ChatBots\ReplayCapture.cs" />
